requirements.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.13.0
pandas==2.1.1
lxml==4.9.3
python-dotenv==1.0.0
currency-converter==0.5.4
Pillow==10.0.0
openai==1.3.7
